+++ EvolutionChamber/src/java/sc2/asset/SC2AssetType.java
@@ -46,7 +46,7 @@ public class SC2AssetType {
 	final public int prov_f;
 
 	/** supply cost. units have >=0 supply, tech/structs have null supply */
-	final public Integer cost_f;
+	final public Float cost_f;
 	/** cargo size */
 	final public int size;
 	/** cargo capacity */
@@ -59,8 +59,8 @@ public class SC2AssetType {
 	/**
 	** Return the food cost, or 0 if this asset type does not use supply
 	*/
-	public int supply() {
-		return cost_f == null? 0: (int)cost_f;
+	public float supply() {
+		return cost_f == null? 0: (float)cost_f;
 	}
 
 	/**
@@ -97,7 +97,7 @@ public class SC2AssetType {
 		SC2HealthSchema stat_hp, SC2HealthSchema stat_sp, SC2EnergySchema stat_ep,
 		double speed, int range, int sight,
 		EnumSet<Modifier> mods, SC2Attack atk_g, SC2Attack atk_a,
-		int prov_f, Integer cost_f, int size, int cargo
+		int prov_f, Float cost_f, int size, int cargo
 	) {
 		this.name = non_null("name", name);
 		this.race = non_null("race", race);
@@ -192,7 +192,7 @@ public class SC2AssetType {
 			  mods, atk_g, atk_a, prov_f, null, 0, 0);
 		}
 
-		public SC2AssetType buildUnit(int cost_f, int size, int cargo) {
+		public SC2AssetType buildUnit(float cost_f, int size, int cargo) {
 			return new SC2AssetType(name, race, actions,
 			  stat_hp, stat_sp, stat_ep, speed, range, sight,
 			  mods, atk_g, atk_a, prov_f, cost_f, size, cargo);
